+++ frontend/css/results/_DesktopSectionPanel.scss
@@ -7,7 +7,6 @@
 
   &__meetings {
     display: inline-flex;
-    margin-left: 16px;
 
   }
